requirements.txt
pyserial-asyncio>=0.6
pyserial>=3.5

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Jupyter notebook support
ipywidgets>=8.0.0
jupyter>=1.0.0
nest_asyncio>=1.5.0